Исправляет компоненты
